+++ CommonplaceBookApp/DocumentTableController.swift
@@ -16,7 +16,6 @@ protocol DocumentTableControllerDelegate: AnyObject {
 
 /// Given a notebook, this class can manage a table that displays the hashtags and pages of that notebook.
 public final class DocumentTableController: NSObject {
-
   /// Designated initializer.
   public init(
     tableView: UITableView,
@@ -147,7 +146,7 @@ extension DocumentTableController: UITableViewDelegate {
         return
       }
       let textEditViewController = TextEditViewController(
-        parsingRules: notebook.parsingRules
+        notebook: notebook
       )
       textEditViewController.pageIdentifier = viewProperties.pageKey
       textEditViewController.markdown = markdown
@@ -338,10 +337,10 @@ private extension DocumentTableController {
     filteredHashtag: String?,
     filteredPageIdentifiers: Set<String>?
   ) -> Snapshot {
-    let snapshot = Snapshot()
+    var snapshot = Snapshot()
     if !hashtags.isEmpty {
       snapshot.appendSections([.hashtags])
-      snapshot.appendItems(hashtags.map({ Item.hashtag($0) }))
+      snapshot.appendItems(hashtags.map { Item.hashtag($0) })
     }
     snapshot.appendSections([.documents])